Changes to Rendering, Logic and Breakoutroom UIElement
Added a list for update order and render  order in favour of recursive updates. This eliminates a propagating effect of UIElements "lagging behind" their parents.
Also reversed the collision detection update order so that colliders "in front" have an implicitly higher priority.
Extended functionality of the breakoutroom UIElement. This was rather complicated as it requires support for arguments in message passing, better lock policies for colliders, reparenting of UIElements at runtime and deletion of UIElements at runtime.
The current implementation of collider locks is not side effect free. The collision system deadlocks if its UIElement is deleted while it holds the lock during OnClickEnter and OnClickUpdate.